Two bugs fixed: - WRITE within the pseudoFS did not return NFS4ERR_ROFS - Bad XDR encoding in function NFS4_FSALattr_To_Fattr . This concerned attributes FATTR4_TIME_MODIFY_SET and FATTR4+TIME_ACCESS_SET that are used only by FSAL_PROXY
Useless checksum has been removed from NFSv4 filehandles.

+++ src/Cache_inode/cache_inode_open_close.c
@@ -142,7 +142,7 @@ cache_inode_status_t cache_inode_open(cache_entry_t * pentry,
         }
 
       pentry->object.file.open_fd.fileno =
-          (int)FSAL_FILENO(&(pentry->object.file.open_fd.fd));
+          FSAL_FILENO(&(pentry->object.file.open_fd.fd));
       pentry->object.file.open_fd.openflags = openflags;
 
 #ifdef _DEBUG_CACHE_INODE
